change unit testing framework to jest, seperate classes into seperate files
