--- mocha.opts
@@ -1,3 +1,0 @@
---require ts-node/register
-test/mocha-globals.js
-src/**/*.spec.ts
