+++ tsconfig.json
@@ -4,6 +4,10 @@
     "rootDir": "src",
     "outDir": "dist",
     "target" : "es5",
+    "lib": [
+      "es6",
+      "dom"
+    ],
     "sourceMap" : true,
     "declaration": true,
     "declarationDir": "dist/types"
